Cleaned up code

+++ src/exercises/Rectangle.java
@@ -6,18 +6,18 @@ public class Rectangle {
     public static void main(String[] args) {
         // write code in here that you want to execute
         // declare variables
-        int height;
-        int width;
-        int area;
+        double height;
+        double width;
+        double area;
         Scanner in;
 
         // get  height and width from user
         in = new Scanner(System.in);
         System.out.println("Enter the height of your rectangle: ");
-        height = in.nextInt();
+        height = in.nextDouble();
 
         System.out.println("Enter the width of your rectangle: ");
-        width = in.nextInt();
+        width = in.nextDouble();
 
         // calculate area
         area = width * height;
